Move GenerateContent to llms package
