+++ cmd/main.go
@@ -6,11 +6,11 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
 
 	"github.com/darmenliu/nuwa-terminal-chat/pkg/agents"
 	"github.com/darmenliu/nuwa-terminal-chat/pkg/cmdexe"
+	"github.com/darmenliu/nuwa-terminal-chat/pkg/llms"
 	"github.com/darmenliu/nuwa-terminal-chat/pkg/parser"
 	"github.com/darmenliu/nuwa-terminal-chat/pkg/prompts"
 
@@ -21,10 +21,6 @@ import (
 	lcagents "github.com/tmc/langchaingo/agents"
 	"github.com/tmc/langchaingo/chains"
 	lcllms "github.com/tmc/langchaingo/llms"
-	"github.com/tmc/langchaingo/llms/anthropic"
-	"github.com/tmc/langchaingo/llms/googleai"
-	lcollama "github.com/tmc/langchaingo/llms/ollama"
-	"github.com/tmc/langchaingo/llms/openai"
 	"github.com/tmc/langchaingo/tools"
 )
 
@@ -106,95 +102,6 @@ func GetPromptAccordingToCurrentMode(current string, in string) string {
 	return sysPrompt + "\n" + in
 }
 
-func GetLLMBackend(ctx context.Context) (lcllms.Model, error) {
-	logger := pterm.DefaultLogger.WithLevel(pterm.LogLevelTrace)
-	llmBackend := os.Getenv("LLM_BACKEND")
-	modelName := os.Getenv("LLM_MODEL_NAME")
-	apiKey := os.Getenv("LLM_API_KEY")
-
-	if llmBackend == "" {
-		llmBackend = "gemini"
-		modelName = "gemini-1.5-pro"
-	}
-
-	if apiKey == "" {
-		logger.Error("LLM_API_KEY is not set")
-		return nil, errors.New("LLM_API_KEY is not set")
-	}
-
-	var model lcllms.Model
-	var err error
-	switch llmBackend {
-	case "gemini":
-		model, err = googleai.New(ctx, googleai.WithAPIKey(apiKey), googleai.WithDefaultModel(modelName))
-	case "ollama":
-		serverUrl := os.Getenv("OLLAMA_SERVER_URL")
-		if serverUrl == "" {
-			logger.Error("OLLAMA_SERVER_URL is not set")
-			return nil, errors.New("OLLAMA_SERVER_URL is not set")
-		}
-		model, err = lcollama.New(lcollama.WithModel(modelName), lcollama.WithServerURL(serverUrl))
-	case "groq":
-		model, err = openai.New(
-			openai.WithModel("llama3-8b-8192"),
-			openai.WithBaseURL("https://api.groq.com/openai/v1"),
-			openai.WithToken(apiKey),
-		)
-	case "deepseek":
-		baseurl := os.Getenv("LLM_BASE_URL")
-		if baseurl == "" {
-			logger.Error("LLM_BASE_URL is not set")
-			return nil, errors.New("LLM_BASE_URL is not set")
-		}
-		model, err = openai.New(
-			openai.WithModel(modelName),
-			openai.WithBaseURL(baseurl),
-			openai.WithToken(apiKey),
-		)
-	case "claude":
-		model, err = anthropic.New(
-			anthropic.WithModel("claude-3-5-sonnet-20240620"),
-			anthropic.WithToken(apiKey),
-		)
-	default:
-		return nil, fmt.Errorf("unknown LLM backend: %s", llmBackend)
-	}
-
-	if err != nil {
-		logger.Error(fmt.Sprintf("failed to create %s client, error:", llmBackend), logger.Args("err", err.Error()))
-		return nil, err
-	}
-
-	return model, nil
-}
-
-// GenerateContent generates content using the specified prompt.
-// It takes a context.Context object and a prompt string as input.
-// The function returns the generated content as a string and an error if any.
-// The function first reads the values of environment variables LLM_BACKEND, LLM_MODEL_NAME, and LLM_TEMPERATURE.
-// If LLM_BACKEND is empty, it defaults to "gemini" and LLM_MODEL_NAME defaults to "gemini-1.5-pro".
-// The function then creates a model based on the value of LLM_BACKEND using the specified context, model name, and temperature.
-// The model is closed using the CloseBackend method before returning the generated content.
-// If an error occurs during model creation or content generation, an error is returned.
-func GenerateContent(ctx context.Context, prompt string) (string, error) {
-
-	model, err := GetLLMBackend(ctx)
-	if err != nil {
-		return "", fmt.Errorf("failed to get LLM backend: %w", err)
-	}
-
-	modeTemperature, err := strconv.ParseFloat(os.Getenv("LLM_TEMPERATURE"), 64)
-	if err != nil {
-		return "", fmt.Errorf("failed to parse LLM_TEMPERATURE: %w", err)
-	}
-
-	resp, err := lcllms.GenerateFromSinglePrompt(ctx, model, prompt, lcllms.WithTemperature(modeTemperature))
-	if err != nil {
-		return "", fmt.Errorf("failed to generate content: %w", err)
-	}
-	return resp, nil
-}
-
 type NuwaChat struct {
 	model        lcllms.Model
 	chatHistory  []lcllms.MessageContent
@@ -202,7 +109,7 @@ type NuwaChat struct {
 }
 
 func NewNuwaChat(ctx context.Context, systemPrompt string) (*NuwaChat, error) {
-	model, err := GetLLMBackend(ctx)
+	model, err := llms.GetLLMBackend(ctx)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get LLM backend: %w", err)
 	}
@@ -341,7 +248,7 @@ func handleChatMode(ctx context.Context, input string) error {
 func handleCmdMode(ctx context.Context, prompt string) error {
 	logger := pterm.DefaultLogger.WithLevel(pterm.LogLevelTrace)
 
-	rsp, err := GenerateContent(ctx, prompt)
+	rsp, err := llms.GenerateContent(ctx, prompt)
 	if err != nil {
 		logger.Error("NUWA TERMINAL: failed to generate content,", logger.Args("err", err.Error()))
 		return err
@@ -416,7 +323,7 @@ func handleNuwaScript(ctx context.Context, filepath string) error {
 	}
 	prompt = prompt + "\n" + scriptPrompt
 	// generate content
-	rsp, err := GenerateContent(ctx, prompt)
+	rsp, err := llms.GenerateContent(ctx, prompt)
 	if err != nil {
 		logger.Error("NUWA TERMINAL: failed to generate content,", logger.Args("err", err.Error()))
 		return err
@@ -435,7 +342,7 @@ func handleNuwaScript(ctx context.Context, filepath string) error {
 func handleTaskMode(ctx context.Context, prompt string) error {
 	logger := pterm.DefaultLogger.WithLevel(pterm.LogLevelTrace)
 
-	rsp, err := GenerateContent(ctx, prompt)
+	rsp, err := llms.GenerateContent(ctx, prompt)
 	if err != nil {
 		logger.Error("NUWA TERMINAL: failed to generate content,", logger.Args("err", err.Error()))
 		return err
@@ -511,7 +418,7 @@ func prepareScriptFile(filename, content string) (string, error) {
 func handleAgentMode(ctx context.Context, input string) error {
 	logger := pterm.DefaultLogger.WithLevel(pterm.LogLevelTrace)
 
-	llm, err := GetLLMBackend(ctx)
+	llm, err := llms.GetLLMBackend(ctx)
 	if err != nil {
 		logger.Error("NUWA TERMINAL: failed to get LLM backend,", logger.Args("err", err.Error()))
 		return err
